+++ com/ab/papercrafts/Editor/ForceBootSceneOnPlay/ForceBootFirstBuildSceneOnPlay.cs
@@ -1,4 +1,3 @@
-#if UNITY_EDITOR
 using System;
 using UnityEditor;
 using UnityEditor.SceneManagement;
@@ -97,4 +96,3 @@ namespace com.ab.papercrafts
         }
     }
 }
-#endif